requirements.txt
chardet==2.3.0
decorator==4.0.10
dj-database-url==0.4.2
Django==1.9.5
django-filter==0.13.0
django-mysql==1.0.9
django-sslify==0.2.7
djangorestframework==3.4.0
html5lib==1.0b8
iso8601==0.1.11
isodate==0.5.4
keepalive==0.5
Markdown==2.6.6
mysql==0.0.1
MySQL-python==1.2.5
networkx==1.11
nltk==3.2.1
numpy==1.11.0
oauthlib==2.0.2
psycopg2==2.6.2
PyJWT==1.5.0
pyparsing==2.1.5
python-magic==0.4.6
python-openid==2.2.5
python-social-auth==0.3.6
rdflib==4.2.1
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
social-auth-app-django==1.1.0
social-auth-core==1.2.0
SPARQLWrapper==1.7.6
tethne==0.8.1.dev5
Unidecode==0.4.19
wheel==0.24.0
